+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v4.5.2
+* Fixed: Accumulo delete extended data not removing rows 
 * Changed: Elasticsearch to not use locks when flushing it's internal queue
 
 # v4.5.1